all: drop support for Go 1.16 and Go 1.17
