+++ BUILDING.md
@@ -18,7 +18,7 @@ tarball. If you want to help with development of TinyGo itself, you should follo
 LLVM, Clang and LLD are quite light on dependencies, requiring only standard
 build tools to be built. Go is of course necessary to build TinyGo itself.
 
-  * Go (1.16+)
+  * Go (1.18+)
   * Standard build tools (gcc/clang)
   * git
   * CMake
