+++ loader/loader.go
@@ -27,8 +27,6 @@ import (
 	"github.com/tinygo-org/tinygo/goenv"
 )
 
-var addInstances func(*types.Info)
-
 // Program holds all packages and some metadata about the program as a whole.
 type Program struct {
 	config       *compileopts.Config
@@ -158,15 +156,13 @@ func Load(config *compileopts.Config, inputPkg string, clangHeaders string, type
 			EmbedGlobals: make(map[string][]*EmbedFile),
 			info: types.Info{
 				Types:      make(map[ast.Expr]types.TypeAndValue),
+				Instances:  make(map[*ast.Ident]types.Instance),
 				Defs:       make(map[*ast.Ident]types.Object),
 				Uses:       make(map[*ast.Ident]types.Object),
 				Implicits:  make(map[ast.Node]types.Object),
 				Scopes:     make(map[ast.Node]*types.Scope),
 				Selections: make(map[*ast.SelectorExpr]*types.Selection),
 			},
-		}
-		if addInstances != nil {
-			addInstances(&pkg.info)
 		}
 		err := decoder.Decode(&pkg.PackageJSON)
 		if err != nil {
