--- src/internal/itoa/README.md
@@ -1,2 +1,0 @@
-internal/itoa is new to go as of 1.17.
-This directory should be removed when tinygo drops support for go 1.16.
