+++ src/runtime/timer.go
@@ -1,6 +1,3 @@
-//go:build go1.18
-// +build go1.18
-
 // Portions copyright 2009 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
