--- testdata/go1.17.txt
@@ -1,3 +1,0 @@
-slice to array pointer: 1 -2 20 4
-unsafe.Add array: 1 5 8 4
-unsafe.Slice array: 3 3 9 15 4
