+++ testdata/slice.txt
@@ -13,3 +13,6 @@ grow: len=7 cap=8 data: 42 -1 -2 1 2 4 5
 grow: len=7 cap=8 data: 42 -1 -2 1 2 4 5
 grow: len=14 cap=16 data: 42 -1 -2 1 2 4 5 42 -1 -2 1 2 4 5
 bytes: len=6 cap=6 data: 1 2 3 102 111 111
+slice to array pointer: 1 -2 20 4
+unsafe.Add array: 1 5 8 4
+unsafe.Slice array: 3 3 9 15 4
