--- testdata/testing_go118.txt
@@ -1,21 +1,0 @@
---- FAIL: TestBar (0.00s)
-    log Bar
-    log g
-        h
-        i
-        
-    --- FAIL: TestBar/Bar2 (0.00s)
-        log Bar2
-            a
-            b
-            c
-        failed
-        after failed
-    log Bar end
---- FAIL: TestAllLowercase (0.00s)
-    --- FAIL: TestAllLowercase/BETA (0.00s)
-        expected lowercase name, got BETA
-    --- FAIL: TestAllLowercase/BELTA (0.00s)
-        expected lowercase name, got BELTA
-FAIL
-exitcode: 1
