+++ tests/text/template/smoke/smoke_test.go
@@ -1,6 +1,3 @@
-//go:build go1.16
-// +build go1.16
-
 // Copyright 2011 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
